requirements.txt
Flask
openaq
flask_sqlalchemy
python-dotenv